Conflicts:
	pipeline/config/dev-dvillafuerte/test-hibernate.cfg.xml
	pipeline/config/dev-dvillafuerte/testDatabaseConfig.properties
	src/main/java/org/generationcp/middleware/domain/etl/Workbook.java
	src/main/java/org/generationcp/middleware/domain/oms/TermId.java
	src/main/java/org/generationcp/middleware/manager/GdmsTable.java
	src/main/java/org/generationcp/middleware/manager/GenotypicDataManagerImpl.java
	src/main/java/org/generationcp/middleware/operation/saver/WorkbookSaver.java
	src/main/java/org/generationcp/middleware/pojos/gdms/MappingPop.java
	src/main/java/org/generationcp/middleware/pojos/ims/Lot.java
	src/main/java/org/generationcp/middleware/service/FieldbookServiceImpl.java
	src/test/java/org/generationcp/middleware/manager/test/TestGenotypicDataManagerImplUploadFunctions.java
	src/test/resources/test-hibernate.cfg.xml

+++ src/main/java/org/generationcp/middleware/pojos/report/LotReportRow.java
@@ -25,7 +25,7 @@ public class LotReportRow implements Serializable{
 
     private Integer lotId;
     private Integer entityIdOfLot;
-    private Long actualLotBalance;
+    private Double actualLotBalance;
     private Location locationOfLot;
     private Term scaleOfLot;
     private String commentOfLot;
@@ -46,11 +46,11 @@ public class LotReportRow implements Serializable{
         this.entityIdOfLot = entityIdOfLot;
     }
 
-    public Long getActualLotBalance() {
+    public Double getActualLotBalance() {
         return actualLotBalance;
     }
 
-    public void setActualLotBalance(Long actualLotBalance) {
+    public void setActualLotBalance(Double actualLotBalance) {
         this.actualLotBalance = actualLotBalance;
     }
